requirements.txt
fastapi
uvicorn
opencv-python
Pillow
requests
python-multipart
python-dotenv
